Fix all scala error

+++ src/main/scala/cn/academy/core/client/ui/TechUI.scala
@@ -415,7 +415,7 @@ object TechUI {
       }
 
       def blank(ht: Double) = {
-        elemY += ht
+        elemY += ht.toFloat
 
         this
       }
